server/requirements.txt
fastapi
uvicorn
transformers
torch
spotipy
pillow
python-dotenv
